[IMPORTANT] Reorganize totally. Plsease read below:
* Change the directory structure.
    - `bibi` <- formerly `bib/i`
    - `bibi-bookshelf` <- formerly `bib/bookshelf`
* Change the base of relative paths defined in preset files to the preset file itself (formerly the base was `bib/i/index.html`).
* Add `__dist` for distribution/production and `__src` for development.
* Move all the components of application including HTMLs and preset files to `__src`, and add them to the watch target to make them copied/compiled to `__dist` automatically when they are edited.

Developers have to edit files in `__src` not in `__dist`, even if the file you want to edit is a HTML or a preset file, and generate files for `__dist` by running scripts in terminal. Please don't edit files in `__dist` directly, because they are overwritten by the generated files from `__src` (except the contents of the bookshelf folder).